+++ src/TheRestOfYourLife/pdf.h
@@ -29,12 +29,12 @@ class cosine_pdf : public pdf {
   public:
     cosine_pdf(const vec3& w) { uvw.build_from_w(w); }
 
-    virtual double value(const vec3& direction) const override {
+    double value(const vec3& direction) const override {
         auto cosine = dot(unit_vector(direction), uvw.w());
         return (cosine <= 0) ? 0 : cosine/pi;
     }
 
-    virtual vec3 generate() const override {
+    vec3 generate() const override {
         return uvw.local(random_cosine_direction());
     }
 
@@ -47,11 +47,11 @@ class hittable_pdf : public pdf {
   public:
     hittable_pdf(shared_ptr<hittable> p, const point3& origin) : ptr(p), o(origin) {}
 
-    virtual double value(const vec3& direction) const override {
+    double value(const vec3& direction) const override {
         return ptr->pdf_value(o, direction);
     }
 
-    virtual vec3 generate() const override {
+    vec3 generate() const override {
         return ptr->random(o);
     }
 
@@ -68,11 +68,11 @@ class mixture_pdf : public pdf {
         p[1] = p1;
     }
 
-    virtual double value(const vec3& direction) const override {
+    double value(const vec3& direction) const override {
         return 0.5 * p[0]->value(direction) + 0.5 *p[1]->value(direction);
     }
 
-    virtual vec3 generate() const override {
+    vec3 generate() const override {
         if (random_double() < 0.5)
             return p[0]->generate();
         else